refactor(trusty-common): extract DOC-28 catch-up engine behind `catchup` feature; re-wire trusty-mpm (PR1, refs #1762)
Move all 7 catch-up source files from trusty-mpm/src/core/ into
trusty-common/src/catchup/ behind a new `catchup` feature (pulls in
rusqlite as an optional dep). Re-wire trusty-mpm via thin pub-use shims
so all existing crate::core::{catchup,native_session_finder,
claude_mpm_session,claude_mpm_registry}::* call sites continue to compile
without change. Tests move with the code and run under
`cargo test -p trusty-common --features catchup`. The catchup feature is
opt-in — default `cargo check -p trusty-common` does not compile rusqlite.
